requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiofiles>=23.2.0
aiosqlite>=0.20.0
pydantic>=2.0
openpyxl>=3.1.0
